feat: move cli code into discrete package and update imports; create Dockerfile
